ingress_server/services/extractor_service/pyproject.toml
[project]
name = "extractor-service"
description = "Airflow extractor"
version = "0.1.0"
requires-python = ">=3.11"
authors = [
  { name = "Jan Brezina", email = "jan.brezina@tul.cz" },
  { name = "Stepan Moc", email = "stepan.mocik@gmail.com" },
]
maintainers = [{ name = "Stepan Moc", email = "stepan.mocik@gmail.com" }]
license = "BSD-3-Clause"
keywords = ["Airflow", "Zarr", "s3"]

dependencies = [
  "pydantic==2.12.3",
  "python-dotenv==1.2.1",
  "pandas==2.1.2",
  "zarr_fuse @ git+https://github.com/GeoMop/zarr_fuse.git@main",
  "h5py==3.15.1",
]


[project.optional-dependencies]
apache = ["apache-airflow==3.1.1"]

[tool.setuptools.packages.find]
where = ["."]
include = ["extractor_service*"]

[tool.setuptools.package-dir]
"" = "."

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
